# perf(gatsby): cache `createNodeId` locally