This PR will cache the `uuid` calls when you call `createNodeId` repetitively with the same id/namespace.

This saves a large chunk at scale, especially for plugins like Contentful which call this API many times with the same key. Contentful tends to have a lot of nodes and a site that generates 1.6M nodes saves about 60 seconds just for doing this. Most sites won't see much difference.

It also caches the `uuid.v5` call for the namespace, which is often the same.

This is a fork of https://github.com/gatsbyjs/gatsby/pull/27661

> gabe-fs-text (128k pages)
```
success source and transform nodes - 56.154s
success createPages - 28.945s
info bootstrap finished - 89.735s
info Done building in 267.619 sec
```
```
success source and transform nodes - 52.153s
success createPages - 28.687s
info bootstrap finished - 85.447s
info Done building in 262.359 sec
```

> gabe-fs-markdown (128k pages)
```
success source and transform nodes - 65.534s
success createPages - 30.129s
info bootstrap finished - 100.817s
info Done building in 474.506 sec
```
```
success source and transform nodes - 61.841s
success createPages - 29.876s
info bootstrap finished - 96.869s
info Done building in 481.904 sec
```

1.6M node, 30k page Contentful site (Contentful plugin calls createNodeId very often with same args):
```
success Contentful: Create nodes - 206.140s
success source and transform nodes - 215.330s
success createPages - 57.998s
info bootstrap finished - 314.033s
```
```
success source and transform nodes - 151.334s
success Contentful: Create nodes - 142.157s
success createPages - 53.569s
info bootstrap finished - 244.923s
```